requirements.txt
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
matplotlib
keras
tensorflow
keras-resnet
pillow
numpy
pandas_ml
pandas
sklearn
seaborn
scikit-image